challenge-1b/requirements.txt
PyMuPDF==1.22.0             # For extracting text from PDFs
sentence-transformers==2.2.2  # For embeddings (MiniLM etc.)
faiss-cpu==1.7.4             # For vector similarity search
numpy==1.24.3                # Used for handling embeddings